requirements.txt
Django
django-mptt==0.9.0
FeinCMS==1.13.6
Pillow==5.0.0
requests==2.18.4
beautifulsoup4==4.6.0.
